requirements.txt
keras==2.13.1
librosa==0.10.0.post2
tensorflow-cpu==2.13.0
pydub
scipy